re-initial commit
I start from scratch using Android 2.0. As such I created 2 project and will port most of my code which is not much anyway. We will get more traction on this while working on several projects which will use the REST provider as backend. By the end of this year we should have a nice lib...