Back on the RESTProvider - adding linked in exemple with oauth and xml cursor
Working on getting an XML cursor. I will probably use XPath for that. Currently it prints the XML to the log. It uses OAuth for signing with signpot. You need to first allow the application via the authorize button and you should be able to query. There might be a need to restart the app in order for the access token to take effect.